requirements.txt
hazm==0.7.0
python-bidi==0.4.2
wordcloud==1.8.1